homeworks/2025/06-best-practices/pyproject.toml
[project]
name = "06-best-practices"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "awscli>=1.40.26",
    "click>=8.2.1",
    "pandas>=2.3.0",
    "polars>=1.31.0",
    "pyarrow>=20.0.0",
    "python-dotenv>=1.1.1",
    "s3fs>=2025.5.1",
    "scikit-learn==1.5.0",
]

[dependency-groups]
dev = [
    "pytest>=8.4.1",
]
